unit: activate /run/systemd/condition-job-failure check
The /run/systemd/condition-job-failure file, if it contains a positive value
from any of the ones that parse_config() in util.c looks for, will turn all
failed conditions from merely skipping the job, to hard job failures.

This is meant to emulate systemd-git (should be in 218 soon)'s new Assert*=
options that do this on a per-unit file basis, whereas this is global but
volatile.